Removed main screen and started refactor to one map activity
The main nav screen is now gone in favor of going directly to the map.  Still have some work to
do to capture all of the functionality that was on the main screen, eg subregion and query.

Started to refactor into one map activity.  This will make the code much more easy to maintain.
Although I now have one map activity there is still work to do to fragment some other activities,
eg the list view such that the map activity does not  have to launch a phone or tablet
activity.  It should be able to just launch one activity with different layout files for the
different screen sizes.