Simple filtering
Adding in simple filter.  Simple filtering is based on a date and a keyword.  Added new activity
(will be displayed as a dialog on large screens to do simple filtering.  From that activity the user
also has the ability to click 'advanced filter' which will take them to the old filter activity.

This is still a work in progress, this commit has the new activities and layout files.